Merge Dev into Master

See merge request dandelion-classic/Dandelion!2

+++ src/main/kotlin/org/dandelion/classic/blocks/ClassicBlocks.kt
@@ -3,301 +3,301 @@ package org.dandelion.classic.blocks
 import org.dandelion.classic.blocks.model.Block
 
 class Air : Block() {
-    override val id: Byte = 0
+    override val id: UShort = 0u
     override val name: String = "Air"
     override val isDefault: Boolean = true
 }
 
 class Stone : Block() {
-    override val id: Byte = 1
+    override val id: UShort = 1u
     override val name: String = "Stone"
     override val isDefault: Boolean = true
 }
 
 class GrassBlock : Block() {
-    override val id: Byte = 2
+    override val id: UShort = 2u
     override val name: String = "Grass Block"
     override val isDefault: Boolean = true
 }
 
 class Dirt : Block() {
-    override val id: Byte = 3
+    override val id: UShort = 3u
     override val name: String = "Dirt"
     override val isDefault: Boolean = true
 }
 
 class Cobblestone : Block() {
-    override val id: Byte = 4
+    override val id: UShort = 4u
     override val name: String = "Cobblestone"
     override val isDefault: Boolean = true
 }
 
 class Planks : Block() {
-    override val id: Byte = 5
+    override val id: UShort = 5u
     override val name: String = "Planks"
     override val isDefault: Boolean = true
 }
 
 class Sapling : Block() {
-    override val id: Byte = 6
+    override val id: UShort = 6u
     override val name: String = "Sapling"
     override val isDefault: Boolean = true
 }
 
 class Bedrock : Block() {
-    override val id: Byte = 7
+    override val id: UShort = 7u
     override val name: String = "Bedrock"
     override val isDefault: Boolean = true
 }
 
 class FlowingWater : Block() {
-    override val id: Byte = 8
+    override val id: UShort = 8u
     override val name: String = "Flowing Water"
     override val isDefault: Boolean = true
 }
 
 class StationaryWater : Block() {
-    override val id: Byte = 9
+    override val id: UShort = 9u
     override val name: String = "Stationary Water"
     override val isDefault: Boolean = true
 }
 
 class FlowingLava : Block() {
-    override val id: Byte = 10
+    override val id: UShort = 10u
     override val name: String = "Flowing Lava"
     override val isDefault: Boolean = true
 }
 
 class StationaryLava : Block() {
-    override val id: Byte = 11
+    override val id: UShort = 11u
     override val name: String = "Stationary Lava"
     override val isDefault: Boolean = true
 }
 
 class Sand : Block() {
-    override val id: Byte = 12
+    override val id: UShort = 12u
     override val name: String = "Sand"
     override val isDefault: Boolean = true
 }
 
 class Gravel : Block() {
-    override val id: Byte = 13
+    override val id: UShort = 13u
     override val name: String = "Gravel"
     override val isDefault: Boolean = true
 }
 
 class GoldOre : Block() {
-    override val id: Byte = 14
+    override val id: UShort = 14u
     override val name: String = "Gold Ore"
     override val isDefault: Boolean = true
 }
 
 class IronOre : Block() {
-    override val id: Byte = 15
+    override val id: UShort = 15u
     override val name: String = "Iron Ore"
     override val isDefault: Boolean = true
 }
 
 class CoalOre : Block() {
-    override val id: Byte = 16
+    override val id: UShort = 16u
     override val name: String = "Coal Ore"
     override val isDefault: Boolean = true
 }
 
 class Wood : Block() {
-    override val id: Byte = 17
+    override val id: UShort = 17u
     override val name: String = "Wood"
     override val isDefault: Boolean = true
 }
 
 class Leaves : Block() {
-    override val id: Byte = 18
+    override val id: UShort = 18u
     override val name: String = "Leaves"
     override val isDefault: Boolean = true
 }
 
 class Sponge : Block() {
-    override val id: Byte = 19
+    override val id: UShort = 19u
     override val name: String = "Sponge"
     override val isDefault: Boolean = true
 }
 
 class Glass : Block() {
-    override val id: Byte = 20
+    override val id: UShort = 20u
     override val name: String = "Glass"
     override val isDefault: Boolean = true
 }
 
 class RedCloth : Block() {
-    override val id: Byte = 21
+    override val id: UShort = 21u
     override val name: String = "Red Cloth"
     override val isDefault: Boolean = true
 }
 
 class OrangeCloth : Block() {
-    override val id: Byte = 22
+    override val id: UShort = 22u
     override val name: String = "Orange Cloth"
     override val isDefault: Boolean = true
 }
 
 class YellowCloth : Block() {
-    override val id: Byte = 23
+    override val id: UShort = 23u
     override val name: String = "Yellow Cloth"
     override val isDefault: Boolean = true
 }
 
 class ChartreuseCloth : Block() {
-    override val id: Byte = 24
+    override val id: UShort = 24u
     override val name: String = "Chartreuse Cloth"
     override val isDefault: Boolean = true
 }
 
 class GreenCloth : Block() {
-    override val id: Byte = 25
+    override val id: UShort = 25u
     override val name: String = "Green Cloth"
     override val isDefault: Boolean = true
 }
 
 class SpringGreenCloth : Block() {
-    override val id: Byte = 26
+    override val id: UShort = 26u
     override val name: String = "Spring Green Cloth"
     override val isDefault: Boolean = true
 }
 
 class CyanCloth : Block() {
-    override val id: Byte = 27
+    override val id: UShort = 27u
     override val name: String = "Cyan Cloth"
     override val isDefault: Boolean = true
 }
 
 class CapriCloth : Block() {
-    override val id: Byte = 28
+    override val id: UShort = 28u
     override val name: String = "Capri Cloth"
     override val isDefault: Boolean = true
 }
 
 class UltramarineCloth : Block() {
-    override val id: Byte = 29
+    override val id: UShort = 29u
     override val name: String = "Ultramarine Cloth"
     override val isDefault: Boolean = true
 }
 
 class VioletCloth : Block() {
-    override val id: Byte = 30
+    override val id: UShort = 30u
     override val name: String = "Violet Cloth"
     override val isDefault: Boolean = true
 }
 
 class PurpleCloth : Block() {
-    override val id: Byte = 31
+    override val id: UShort = 31u
     override val name: String = "Purple Cloth"
     override val isDefault: Boolean = true
 }
 
 class MagentaCloth : Block() {
-    override val id: Byte = 32
+    override val id: UShort = 32u
     override val name: String = "Magenta Cloth"
     override val isDefault: Boolean = true
 }
 
 class RoseCloth : Block() {
-    override val id: Byte = 33
+    override val id: UShort = 33u
     override val name: String = "Rose Cloth"
     override val isDefault: Boolean = true
 }
 
 class DarkGrayCloth : Block() {
-    override val id: Byte = 34
+    override val id: UShort = 34u
     override val name: String = "Dark Gray Cloth"
     override val isDefault: Boolean = true
 }
 
 class LightGrayCloth : Block() {
-    override val id: Byte = 35
+    override val id: UShort = 35u
     override val name: String = "Light Gray Cloth"
     override val isDefault: Boolean = true
 }
 
 class WhiteCloth : Block() {
-    override val id: Byte = 36
+    override val id: UShort = 36u
     override val name: String = "White Cloth"
     override val isDefault: Boolean = true
 }
 
 class Flower : Block() {
-    override val id: Byte = 37
+    override val id: UShort = 37u
     override val name: String = "Flower"
     override val isDefault: Boolean = true
 }
 
 class Rose : Block() {
-    override val id: Byte = 38
+    override val id: UShort = 38u
     override val name: String = "Rose"
     override val isDefault: Boolean = true
 }
 
 class BrownMushroom : Block() {
-    override val id: Byte = 39
+    override val id: UShort = 39u
     override val name: String = "Brown Mushroom"
     override val isDefault: Boolean = true
 }
 
 class RedMushroom : Block() {
-    override val id: Byte = 40
+    override val id: UShort = 40u
     override val name: String = "Red Mushroom"
     override val isDefault: Boolean = true
 }
 
 class BlockOfGold : Block() {
-    override val id: Byte = 41
+    override val id: UShort = 41u
     override val name: String = "Block of Gold"
     override val isDefault: Boolean = true
 }
 
 class BlockOfIron : Block() {
-    override val id: Byte = 42
+    override val id: UShort = 42u
     override val name: String = "Block of Iron"
     override val isDefault: Boolean = true
 }
 
 class DoubleSlab : Block() {
-    override val id: Byte = 43
+    override val id: UShort = 43u
     override val name: String = "Double Slab"
     override val isDefault: Boolean = true
 }
 
 class Slab : Block() {
-    override val id: Byte = 44
+    override val id: UShort = 44u
     override val name: String = "Slab"
     override val isDefault: Boolean = true
 }
 
 class Bricks : Block() {
-    override val id: Byte = 45
+    override val id: UShort = 45u
     override val name: String = "Bricks"
     override val isDefault: Boolean = true
 }
 
 class TNT : Block() {
-    override val id: Byte = 46
+    override val id: UShort = 46u
     override val name: String = "TNT"
     override val isDefault: Boolean = true
 }
 
 class Bookshelf : Block() {
-    override val id: Byte = 47
+    override val id: UShort = 47u
     override val name: String = "Bookshelf"
     override val isDefault: Boolean = true
 }
 
 class MossyCobblestone : Block() {
-    override val id: Byte = 48
+    override val id: UShort = 48u
     override val name: String = "Mossy Cobblestone"
     override val isDefault: Boolean = true
 }
 
 class Obsidian : Block() {
-    override val id: Byte = 49
+    override val id: UShort = 49u
     override val name: String = "Obsidian"
     override val isDefault: Boolean = true
 }
